% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Title}
\format{
a data frame with x rows and x variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{

}
\usage{
key_crop_yields
}
\description{
Title
}
\section{Background}{

xxx The data was collected fromxxx
}

\examples{
print(key_crop_yields)
}
\keyword{datasets}
